% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_add_h_leftright.R
\name{ew_add_h_leftright}
\alias{ew_add_h_leftright}
\title{Add left and right interpolated and extrapolated heights to segments of equal width graph.}
\usage{
ew_add_h_leftright(g)
}
\arguments{
\item{g}{should be an equal width graph, represented as a tibble with
vectors p and h.}
}
\value{
The augmented graph as a tibble.

Examples
g <- ew_add_h_leftright(g)
}
\description{
For a description of equal width probability graphs see above.
}
