% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_add_surface.R
\name{ew_add_surface}
\alias{ew_add_surface}
\title{Add surface and cumulative surface to segments of equal width graph.}
\usage{
ew_add_surface(g)
}
\arguments{
\item{g}{should be an equal width graph, represented as a tibble with
vectors p and h, h_left and h_right.

For a description of equal width probability graphs see ew_minmaxcumh_p.R.}
}
\value{
The augmented graph as a tibble.

Examples
g <- ew_add_surface(g)
}
\description{
This also means adapting the columns h, h_left and h_right
in order to normalize the total chance mass to 1.
}
