% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_d.R
\name{ew_d}
\alias{ew_d}
\title{Probability density function for ew_graph}
\usage{
ew_d(g, p)
}
\arguments{
\item{g}{An ew_graph.}

\item{p}{A real value from [0, 1].}
}
\value{
The (approximated) chance density for p on g.
}
\description{
Approximate the density of an ew_graph anywhere on the
line [0, 1]. In other words return the (approximated) chance density of the
ew_graph g for any value of p on [0, 1].
}
