% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_from_vec.R
\name{ew_from_vec}
\alias{ew_from_vec}
\title{Create an equal width probability graph from vector \code{v}}
\usage{
ew_from_vec(v)
}
\arguments{
\item{v}{The input vector.}
}
\value{
An equal width chance graph representation of the given vector.
}
\description{
We assume that the elements of the vector
represent the probabilities p, of 1/S - 1/2S, ..., S/S - 1/2S.
\code{S} is the number of elements, so the length, of the vector.
We call the corresponding elements of the equal width probability
graph \emph{segments} of the probability graph.
}
