% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{ew_get_cumsurface}
\alias{ew_get_cumsurface}
\title{Get the cumulative surface column of an equal width graph.}
\usage{
ew_get_cumsurface(g)
}
\arguments{
\item{g}{The equal width graph}
}
\value{
The cumulative surface column as a vector
}
\description{
Get the cumulative surface column of an equal width graph.
}
\examples{
  g <- ew_from_vec(c(1, 2, 3))
  ew_get_cumsurface(g)
}
