% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{ew_get_h_left}
\alias{ew_get_h_left}
\title{Get the h_left column of an equal width graph.}
\usage{
ew_get_h_left(g)
}
\arguments{
\item{g}{The equal width graph}
}
\value{
The h_left column as a vector
}
\description{
Get the h_left column of an equal width graph.
}
\examples{
  g <- ew_from_vec(c(1, 2, 3))
  ew_get_h_left(g)
}
