% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_halfsegment_cumh_p.R
\name{ew_halfsegment_cumh_p}
\alias{ew_halfsegment_cumh_p}
\title{In ew half segment, given cumulative value for h, find corresponding p value}
\usage{
ew_halfsegment_cumh_p(h1, h2, sd, U)
}
\arguments{
\item{h1}{The height of the left side of the trapezium.}

\item{h2}{The height of the right side of the trapezium.}

\item{sd}{Alias for surface_delta:
The left part of the surface of the trapezium
bordered by an imaginary line perpendicular to the
base line of the trapezium.
sd > 0, and smaller than the surface of the
trapezium.}

\item{U}{The length of the baseline of the trapezium.}
}
\value{
The computed value of x, which is the point where the imaginary line
crosses the base line.
}
\description{
An ew half segment of an equal width graph, or ewgraph,
is a rectangular trapezium with
baseline size U and two parallel sides perpendicular
to the base line.
}
\examples{
  x <- ew_halfsegment_cumh_p(h1 = 1, h2 = 0, sd = 0.0125, U = 0.25)
}
