% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_high_density_interval.R
\name{ew_high_density_interval}
\alias{ew_high_density_interval}
\title{highest part of the graph with a given cumulative chance density}
\usage{
ew_high_density_interval(g, high_density_area = 0.999)
}
\arguments{
\item{g}{An ewgraph.}

\item{high_density_area}{When 1 return \code{g} as is.
When smaller than 1, the highest part of the ewgraph \code{g}
will be shown
which has a cumulative chance density equal to \code{high_density_interval}
The idea is that this cuts away the uninteresting low value left
and/or right part of the graph,
and leaves only the rest.}
}
\value{
A vector of the form
\code{c(left = i_left, right = i_right)}.
Here \code{i_left} and \code{i_right} are respectively
the leftmost and rightmost ew segment of the HDI.
}
\description{
highest part of the graph with a given cumulative chance density
}
\examples{
  S <- 1000
  g <- ew_from_vec(dbinom(30, 300, partition_0_1(S)))
  lr  <- ew_high_density_interval(g)
  print(lr)
}
