% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_minmaxcumh_p.R
\name{ew_maxcumh_p}
\alias{ew_maxcumh_p}
\title{maximum error rate for given certainty}
\usage{
ew_maxcumh_p(g, cert = 0.95)
}
\arguments{
\item{g}{An ewgraph.}

\item{cert}{The certainty. Lies between 0 and 1.}
}
\value{
The maximum error rate of \code{g}, given certainty \code{cert}.
}
\description{
Returns the lowest (possibly interpolated)
error rate, p, for which the accumulation
of the chance density, h, starting from the first element of g and going to
the last element of \code{g} is \verb{>= cert}.
}
\examples{
  S <- 1000
  g <- ew_from_vec(dbinom(30, 300, partition_0_1(S)))
  max_cumh_p <- ew_maxcumh_p(g)
  print(max_cumh_p)
}
