% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_maxh.R
\name{ew_maxh}
\alias{ew_maxh}
\title{Return the (possibly interpolated) highest value of h, h_left, or h_right
and corresponding error rate, p.}
\usage{
ew_maxh(g)
}
\arguments{
\item{g}{An ewgraph.}
}
\value{
c(h = h_highest, p = p_highest)
}
\description{
Return as a named list of the form c(h = h_highest, p = p_highest).
}
\examples{
  S <- 1000
  g <- ew_from_vec(dbinom(30, 300, partition_0_1(S)))
  hp_max <- ew_maxh(g)
  print(hp_max)
}
