% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_middle_peaked.R
\name{ew_middle_peaked}
\alias{ew_middle_peaked}
\title{Return TRUE iff the highest value of h is reached
neither for the h_left of the first segment
nor for the h_right of the last segment}
\usage{
ew_middle_peaked(g)
}
\arguments{
\item{g}{An ewgraph.}
}
\value{
TRUE or FALSE.
}
\description{
Return TRUE iff the highest value of h is reached
neither for the h_left of the first segment
nor for the h_right of the last segment
}
\examples{
  S <- 1000
  g <- ew_from_vec(dbinom(30, 300, partition_0_1(S)))
  middle_peaked <- ew_middle_peaked(g)
  print(middle_peaked)
}
