% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_minmaxcumh_p.R
\name{ew_mincumh_p}
\alias{ew_mincumh_p}
\title{minimum error rate for given certainty}
\usage{
ew_mincumh_p(g, cert = 0.95)
}
\arguments{
\item{g}{An ewgraph.}

\item{cert}{The certainty. Lies between 0 and 1.}
}
\value{
The minimum error rate of \code{g}, given certainty \code{cert}.
}
\description{
Returns the highest (possibly interpolated) error rate, p, for which the accumulation
of the chance density, h, starting from the last element of \code{g} and going to
the first element of \code{g} is \verb{>= cert}.
}
\examples{
  S <- 1000
  g <- ew_from_vec(dbinom(30, 300, partition_0_1(S)))
  min_cumh_p <- ew_mincumh_p(g)
  print(min_cumh_p)
}
