% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{ew_S}
\alias{ew_S}
\title{Get the number of segments, i.e. number of rows, of an equal width graph.}
\usage{
ew_S(g)
}
\arguments{
\item{g}{The equal width graph}
}
\value{
The number of segments, this is, rows of g
}
\description{
Get the number of segments, i.e. number of rows, of an equal width graph.
}
\examples{
  g <- ew_from_vec(c(1, 2, 3))
  ew_S(g)
}
