% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ew_validate}
\alias{ew_validate}
\title{Check whether object is a valid equal width graph}
\usage{
ew_validate(g, verbose = TRUE)
}
\arguments{
\item{g}{The object to be checked}

\item{verbose}{TRUE or FALSE}
}
\value{
TRUE or FALSE
When g is invalid as an equal width graph
and verbose == TRUE, info why g is invalid
is printed to the console.
}
\description{
Check whether object is a valid equal width graph
}
\examples{
  g <- ew_validate(c(1, 2, 3))
  ew_validate(g)
}
