% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewgraph-package.R
\docType{package}
\name{ewgraph-package}
\alias{ewgraph}
\alias{ewgraph-package}
\title{ewgraph: Equal width representation of chance graphs}
\description{
Chance graphs are represented as tibbles, based on an even spaced base line on (0, 1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cfjdoedens/ewgraph}
  \item \url{https://cfjdoedens.github.io/ewgraph/}
  \item Report bugs at \url{https://github.com/cfjdoedens/ewgraph/issues}
}

}
\author{
\strong{Maintainer}: Crist-Jan Doedens \email{cfjdoedens@chance-matters.nl}

}
\keyword{internal}
