% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{find_non_monotone_pairs}
\alias{find_non_monotone_pairs}
\title{Find input and output pairs of a function that are not monotone rising.}
\usage{
find_non_monotone_pairs(f, in_vals, strictly = FALSE)
}
\arguments{
\item{f}{The function to test. It should be a function that takes a numeric
vector and returns a numeric vector.}

\item{in_vals}{Vector of numeric input values for f.}

\item{strictly}{A boolean. If TRUE, checks for strictly
increasing (f(x2) > f(x1)).
If FALSE (default), checks for non-decreasing (f(x2) >= f(x1)).}
}
\value{
A data frame with columns \code{input1}, \code{output1}, \code{input2}, \code{output2}
containing the offending pairs of input and their outputs.
}
\description{
Gather pairs of input and output in a would be monotone rising
function where the function is not monotone rising.
This function is useful for debugging purposes.
Show the pairs of monotone rising inputs, with their non rising outputs.
}
\examples{
# Example function that is not monotone rising
example_function <- function(x) {
  ifelse(x < 0.5, x, 1 - x)
}

# Find non-monotone pairs in the example function.
x <- find_non_monotone_pairs(example_function, seq(
  from = 0,
  to = 1,
  length.out = 100), strictly = FALSE)
}
