% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{nonnegint}
\alias{nonnegint}
\title{Check whether object is vector of non negative integers}
\usage{
nonnegint(i)
}
\arguments{
\item{i}{The object to check.}
}
\value{
TRUE or FALSE
}
\description{
Check whether object is vector of non negative integers
}
\examples{
  nonnegint(c(0, 1))
}
