% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{partition}
\alias{partition}
\title{Partition line piece on the real line into equally sized consecutive segments}
\usage{
partition(begin = 0, end = 1, S = 1000)
}
\arguments{
\item{begin}{The start of the line piece.}

\item{end}{The end of the line piece.}

\item{S}{An integer >= 1. The number of segments to partition into.}
}
\value{
The vector of segment midpoints.
}
\description{
Each segment is represented by its midpoint.
}
\examples{
  # Returns the vector c(0.1, 0.3, 0.5, 0.7, 0.9).
  partition(S = 5)
}
