% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posint}
\alias{posint}
\title{Check whether object is vector of positive integers}
\usage{
posint(i)
}
\arguments{
\item{i}{The object to check.}
}
\value{
TRUE or FALSE
}
\description{
Check whether object is vector of positive integers
}
\examples{
  posint(c(1, 2))
}
