% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{unity}
\alias{unity}
\title{One argument constant function that returns 1}
\usage{
unity(x)
}
\arguments{
\item{x}{Should be numeric.}
}
\value{
1 for x being numeric, otherwise an error.
}
\description{
The domain of this function is the numeric values. Other input
values will result in an error.
}
\examples{
  # Returns 1.
  unity(-0.034)

  # Returns 1.
  unity(rnorm(1, mean = 10^10, sd = 10^20))

  # Returns 1.
  unity(NaN)

  # Returns 1.
  unity(Inf)

  # Returns c(1, 1).
  unity(c(3, 7))

  # An error is generated for non-numeric input.
  tryCatch({
    unity("a")
  }, error = function(e) {
    cat('unity("a") procuded the following error message:', conditionMessage(e), "\n")
    NA
  })
}
