% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew_utils.R
\name{partition_0_1}
\alias{partition_0_1}
\title{Partition \code{[0, 1]} into equally sized consecutive segments}
\usage{
partition_0_1(S = 1000)
}
\arguments{
\item{S}{An integer >= 1. The number of segments to partition into.}
}
\value{
The vector of segment midpoints.
}
\description{
Each segment is represented by its midpoint.
}
\examples{
  # Returns the vector c(0.1, 0.3, 0.5, 0.7, 0.9).
  partition_0_1(S = 5)
}
